Move parseTemplate function to common testHelpers module
